requirements.txt
PyGithub
pylint
